Move createTransformer to parent class
